[Git] Change pre-commit plugins
- remove flake8 from dependencies
- remove flake8 pre-commit plugin
- add bandit to dependencies
- add bandit pre-commit plugin
- add pylint to dependencies
- add argument for black and isort to read config files

The codebase is not quite there to always run flake8 via a commit hook.
Codacy does also use bandit. Adding it to the pre commit hooks would however help to avoid fix commits or git commit --amend followed by git push --force-with-lease.
Added pylint and pylint-django to Pipfile to be installed in the dev environment to be used during development.