Add field weights to variable search
Matches in the variable names are now most important,
this helps with navigational searches.
With all fields valued equally searched variables were pushed back in
the result list.
The start of the result list was often populated by Variables,
that where merely linked to the intended variable.
This was because they were linked to the variable and contained its
name in other fields, that are searched through.