Limit LabelTable size
Problem
-------

A variable with many value labels and/or many
related variables would cause a server timeout.
This is caused by the generation of a string
containing HTML that is rendered by a LabelTable
instance.

Solution
--------

Stop the LabelTable from rendering if a variable
has more than 100 related variables or more than a
100 value labels.