Disable db constraints for BasketVariable variable
Variable ForeignKey field was already decoupled from
the actual variable through the on_delete=models.DO_NOTHING
setting.
Deleting a variable however still caused django to throw an integrity
error due to the database constraints.

for the System to be able to delete variables without deleting
BasketVariables, db_constraints was set to False for the
variable field of BasketVariable.