Add install and setup of js libs to entrypoint
In production a volume is shared between nginx and web container
to enable nginx to serve the JavaScript files.
This overwrites the JavaScript modules created at image build time.
At container start these modules are now rebuild in the background.

The `npm run build` script was modified to delete JavaScript and css
files before recreating them with webpack.
This is done to prevent the accumulation of files, no longer used.