Fix basket view's remove_concept function
Failure
-------

The function iterated over all variables of a concept and tried to
delete them from the passed basket.
If one of the variables was not in the basket, the function
caused a 404 error.

Solution
--------

Basket variables are now filtered by baket id and variable concept.
This yields only existing BasketVariables, which are then deleted.

Close #1274