Fix 500 error in script generation
When initiating a script generator,
workspace.scripts.soep_datasets.SoepDatasets.get_dict() is called.
The SoepDatasets class is initiated with the content of a json file:
ddionrails/workspace/datafiles.json

If a variable inside a basket does not belong to a dataset
in this json file, a KeyError will be raise at
ddionrails.workspace.scripts.soep_datasets.SoepDatasets.get_dict()
This causes the user to receive a 500 internal error.

* get_dict now returns an empty dict for these datasets
* ddionrails.workspace.mixins.SoepMixin._create_dataset_dict()
  handles the the empty dict by defaulting to "" as default values.

* The empty values will exclude variables from being processed and
  add the to the `NOT PROCESSED` section of the script.

Close #1287